GridLayout: warn at runtime if the row or col is out of bounds
To simplify code generation (by not generating the bounds checking
in the grid_layout_input_for_repeated function), use f64 rather than u16
in GridLayoutInputData, and check bounds in organize_grid_layout.

As a consequence of using f64 to store the values, change the "auto"
magic value from 65535 to 65536. Technically -1 would have worked too
but it's more likely that a binding returns the value -1 at runtime,
and since we're not checking in grid_layout_input_for_repeated, it would
silently trigger auto behavior.

+++ internal/compiler/layout.rs
@@ -278,7 +278,7 @@ impl LayoutConstraints {
 #[derive(Debug, Clone)]
 pub enum RowColExpr {
     Named(NamedReference),
-    Literal(u16),
+    Literal(f64), // u16 range, or ROW_COL_AUTO
 }
 
 /// An element in a GridLayout
